Fix several compiler warnings in C sources.
This takes care of most of the compiler warnings that have crept back into
the C sources.  There are a couple more left, but they will require us to
catch up with current versions of libpng and ffmpeg, so not quite as
trivial as the rest of this patch.

	* modules/ase/ase-metric.c (ase_metric_prnt): Don't try to cast
	XASE_METRIC_DIST to unsigned int, use %p printf mask instead.
	Prevents compiler warning about casting pointer to int of
	different size.

	* src/ui/TTY/terminfo.c (emacs_tparam): cast string to char* to
	keep tparm/ncurses happy.
	Prevents compiler warning.

	* src/ui/X11/glyphs-x.c (x_subwindow_query_geometry): Cast dx, dy
	to int.
	Prevents a compiler warning about wrong types.

	* src/media/sound-alsa.h (sound_alsa_aj_data_s): samplerate is an
	unsigned int.
	Prevents a compiler warning.

+++ src/media/sound-alsa.h
@@ -45,7 +45,7 @@ struct sound_alsa_data_s {
 
 struct sound_alsa_aj_data_s {
 	/* stream props */
-	size_t samplerate;	/* rate used on the device */
+	unsigned int samplerate;	/* rate used on the device */
 	unsigned int channels;	/* channels used on the device */
 	size_t framesize;	/* resulting framesize */
 	snd_pcm_format_t format;
